hejsan

+++ TDA553-lab1-lab2/src/model/model/Car.java
@@ -4,7 +4,7 @@ import java.awt.*;
 
 public abstract class Car extends MotorVehicle {
     
-    public Car(int nrDoors, Color color, int enginePower, String modelName, double X, double Y) {
+    public Car(int nrDoors, Color color, int enginePower, String modelName, int X, int Y) {
         super(nrDoors, color, enginePower, modelName, X, Y);
     }
 }
