+++ TDA553-lab1-lab2/src/model/model/Saab95.java
@@ -4,7 +4,7 @@ import java.awt.*;
 public class Saab95 extends Car {  
     private Turbo turboFunction;
     
-    public Saab95(double X, double Y) {     
+    public Saab95(int X, int Y) {     
         super(4, Color.red, 125, "Saab95", X, Y);              
 	    this.turboFunction = new Turbo(1, 1.3, 125);
         turboFunction.setTurboOff();
