+++ TDA553-lab1-lab2/src/model/model/Truck.java
@@ -5,7 +5,7 @@ public abstract class Truck extends MotorVehicle {
     private Trim trim;
     private IPlatform platform;
         
-    public Truck(int nrDoors, Color color, int enginePower, String modelName, double X, double Y, double trimFactor, IPlatform platform) {
+    public Truck(int nrDoors, Color color, int enginePower, String modelName, int X, int Y, double trimFactor, IPlatform platform) {
         super(nrDoors, color, enginePower, modelName, X, Y);
         this.trim = new Trim(trimFactor, 100);
         this.platform = platform;
