+++ TDA553-lab1-lab2/src/model/model/Volvo240.java
@@ -4,7 +4,7 @@ import java.awt.*;
 public class Volvo240 extends Car  {
     private Trim trim;
     
-    public Volvo240(double x, double y, double trimFactor){
+    public Volvo240(int x, int y, double trimFactor){
         super(4, Color.black, 100, "Volvo240", x, y);
         this.trim = new Trim(trimFactor, 100);
     }
